// src/types/graphql.d.ts
export enum Session_Name_Choices_Enum {
  /** Practice 1 session */
  Practice_1 = 'Practice_1',
  /** Practice 2 session */
  Practice_2 = 'Practice_2',
  /** Practice 3 session */
  Practice_3 = 'Practice_3',
  /** Qualifying session */
  Qualifying = 'Qualifying',
  /** Race session */
  Race = 'Race',
  /** Sprint session */
  Sprint = 'Sprint',
  /** Sprint Qualifying session */
  SprintQualifying = 'Sprint_Qualifying',
  /** Sprint Shootout session */
  SprintShootout = 'Sprint_Shootout',
  /** Test session */
  TestSession = 'Test_Session',
}
// ...
export type GetEventDetailsQuery = {
  __typename?: 'query_root';
  events: Array<{
    __typename?: 'events';
    competition: Array<{
      __typename?: 'sessions';
      scheduled_start_time_utc?: string | null;
      name?: Session_Name_Choices_Enum | null;
      driver_sessions: Array<{
        __typename?: 'driver_sessions';
        driver?: {
          __typename?: 'drivers';
          abbreviation?: string | null;
          full_name?: string | null;
          number?: string | null;
          headshot_url?: string | null;
        } | null;
        constructorByConstructorId?: {
          __typename?: 'constructors';
          color?: string | null;
        } | null;
        results: Array<{
          __typename?: 'results';
          finishing_position?: number | null;
          classified_position?: string | null;
          grid_position?: number | null;
          total_race_time?: bigint | null;
        }>;
        fastest_lap: Array<{
          __typename?: 'laps';
          lap_time?: bigint | null;
          lap_number?: number | null;
        }>;
      }>;
    }>;
    qualifying: Array<{
      __typename?: 'sessions';
      scheduled_start_time_utc?: string | null;
      name?: Session_Name_Choices_Enum | null;
      driver_sessions: Array<{
        __typename?: 'driver_sessions';
        driver?: {
          __typename?: 'drivers';
          abbreviation?: string | null;
          full_name?: string | null;
          number?: string | null;
          headshot_url?: string | null;
        } | null;
        constructorByConstructorId?: {
          __typename?: 'constructors';
          color?: string | null;
        } | null;
        results: Array<{
          __typename?: 'results';
          finishing_position?: number | null;
          q1_time?: bigint | null;
          q2_time?: bigint | null;
          q3_time?: bigint | null;
        }>;
      }>;
    }>;
    practices: Array<{
      __typename?: 'sessions';
      scheduled_start_time_utc?: string | null;
      name?: Session_Name_Choices_Enum | null;
      driver_sessions: Array<{
        __typename?: 'driver_sessions';
        driver?: {
          __typename?: 'drivers';
          abbreviation?: string | null;
          full_name?: string | null;
          number?: string | null;
          headshot_url?: string | null;
        } | null;
        constructorByConstructorId?: {
          __typename?: 'constructors';
          color?: string | null;
        } | null;
        fastest_lap: Array<{
          __typename?: 'laps';
          lap_time?: bigint | null;
          lap_number?: number | null;
        }>;
      }>;
    }>;
  }>;
};
// ...
export type SessionResultsQuery = {
  __typename?: 'query_root';
  sessions: Array<{
    __typename?: 'sessions';
    name?: Session_Name_Choices_Enum | null;
    event?: { __typename?: 'events'; name?: string | null } | null;
    driver_sessions: Array<{
      __typename?: 'driver_sessions';
      constructorByConstructorId?: {
        __typename?: 'constructors';
        name?: string | null;
        color?: string | null;
      } | null;
      driver?: {
        __typename?: 'drivers';
        abbreviation?: string | null;
        full_name?: string | null;
        number?: string | null;
        headshot_url?: string | null;
      } | null;
      results: Array<{
        __typename?: 'results';
        grid_position?: number | null;
        finishing_position?: number | null;
        points?: bigint | number | null;
        status?: string | null;
        classified_position?: string | null;
        total_race_time?: bigint | null;
      }>;
      fastest_lap: Array<{
        __typename?: 'laps';
        lap_number?: number | null;
        stint?: number | null;
        lap_time?: bigint | null;
      }>;
    }>;
  }>;
};